ace-silver/Choice_Coin_Voting/requirements.txt
Flask==2.0.1
flask-mysql==1.5.2
matplotlib==3.4.3
py-algorand-sdk==1.5.0
rsa==4.7.2
tk==0.1.0
Flask-SQLAlchemy==2.5.1
tweepy
python-decouple
tk